fix: subscriber middleware logic
