+++ middleware/subscriber.ts
@@ -1,20 +1,13 @@
 import { storeToRefs } from 'pinia';
-import { get } from '@vueuse/core';
 import { useMainStore } from '~/store';
+import { canBuyNewSubscription } from '~/utils/subscription';
 
 export default defineNuxtRouteMiddleware(async () => {
   const { account } = storeToRefs(useMainStore());
 
-  if (!isDefined(account))
-    return;
+  const canBuy = canBuyNewSubscription(account);
 
-  const { hasActiveSubscription, subscriptions } = get(account);
-  if (hasActiveSubscription) {
-    const renewableSubscriptions = subscriptions.filter(({ actions }) =>
-      actions.includes('renew'),
-    );
-
-    if (!renewableSubscriptions)
-      return navigateTo('/home/subscription');
+  if (!canBuy) {
+    return navigateTo('/home/subscription');
   }
 });
